Replaces useExplorerInfo.tsx with a normal function
