+++ src/components/AppLayout/Header/components/ProviderInfo/ProviderAccessible.tsx
@@ -8,7 +8,7 @@ import Col from 'src/components/layout/Col'
 import Paragraph from 'src/components/layout/Paragraph'
 import WalletIcon from '../WalletIcon'
 import { connected as connectedBg, screenSm, sm } from 'src/theme/variables'
-import { useExplorerInfo } from 'src/logic/hooks/useExplorerInfo'
+import { getExplorerInfo } from 'src/config'
 
 const useStyles = makeStyles({
   network: {
@@ -71,7 +71,7 @@ interface ProviderInfoProps {
 const ProviderInfo = ({ connected, provider, userAddress, network }: ProviderInfoProps): React.ReactElement => {
   const classes = useStyles()
   const addressColor = connected ? 'text' : 'warning'
-  const scanBlockUrl = useExplorerInfo(userAddress)
+  const scanBlockUrl = getExplorerInfo(userAddress)
   return (
     <>
       {!connected && <CircleDot circleSize={35} dotRight={11} dotSize={16} dotTop={24} keySize={14} mode="warning" />}
