+++ src/config/index.ts
@@ -95,3 +95,29 @@ export const buildSafeCreationTxUrl = (safeAddress: string) => {
 
   return `${host}${base}`
 }
+
+export type BlockScanInfo = () => {
+  alt: string
+  url: string
+}
+
+export const getExplorerInfo = (hash?: string | null): BlockScanInfo => {
+  const { name, url } = getNetworkExplorerInfo()
+
+  const blockScanInfo = () => {
+    const result = {
+      url: '',
+      alt: '',
+    }
+    if (!hash) {
+      return result
+    }
+    const type = hash.length > 42 ? 'tx' : 'address'
+    result.url = `${url}${type}/${hash}`
+    result.alt = name || ''
+
+    return result
+  }
+
+  return blockScanInfo
+}
